管理者チェックを追加

+++ src/main/java/actions/EmployeeAction.java
@@ -19,158 +19,183 @@ import services.EmployeeService;
  */
 public class EmployeeAction extends ActionBase {
 
-	private EmployeeService service;
+    private EmployeeService service;
 
-	/**
-	 * メソッドを実行する
-	 */
-	@Override
-	public void process() throws ServletException, IOException {
+    /**
+     * メソッドを実行する
+     */
+    @Override
+    public void process() throws ServletException, IOException {
 
-		service = new EmployeeService();
+        service = new EmployeeService();
 
-		//メソッドを実行
-		invoke();
+        //メソッドを実行
+        invoke();
 
-		service.close();
-	}
+        service.close();
+    }
 
-	/**
-	 * 一覧画面を表示する
-	 * @throws ServletException
-	 * @throws IOException
-	 */
-	public void index() throws ServletException, IOException {
+    /**
+     * 一覧画面を表示する
+     * @throws ServletException
+     * @throws IOException
+     */
+    public void index() throws ServletException, IOException {
 
-		//指定されたページ数の一覧画面に表示するデータを取得
-		int page = getPage();
-		List<EmployeeView> employees = service.getPerPage(page);
+        //管理者かどうかのチェック //追記
+        if (checkAdmin()) { //追記
 
-		//全ての従業員データの件数を取得
-		long employeeCount = service.countAll();
+            //指定されたページ数の一覧画面に表示するデータを取得
+            int page = getPage();
+            List<EmployeeView> employees = service.getPerPage(page);
 
-		putRequestScope(AttributeConst.EMPLOYEES, employees); //取得した従業員データ
-		putRequestScope(AttributeConst.EMP_COUNT, employeeCount); //全ての従業員データの件数
-		putRequestScope(AttributeConst.PAGE, page); //ページ数
-		putRequestScope(AttributeConst.MAX_ROW, JpaConst.ROW_PER_PAGE); //1ページに表示するレコードの数
+            //全ての従業員データの件数を取得
+            long employeeCount = service.countAll();
 
-		//セッションにフラッシュメッセージが設定されている場合はリクエストスコープに移し替え、セッションからは削除する
-		String flush = getSessionScope(AttributeConst.FLUSH);
-		if (flush != null) {
-			putRequestScope(AttributeConst.FLUSH, flush);
-			removeSessionScope(AttributeConst.FLUSH);
-		}
+            putRequestScope(AttributeConst.EMPLOYEES, employees); //取得した従業員データ
+            putRequestScope(AttributeConst.EMP_COUNT, employeeCount); //全ての従業員データの件数
+            putRequestScope(AttributeConst.PAGE, page); //ページ数
+            putRequestScope(AttributeConst.MAX_ROW, JpaConst.ROW_PER_PAGE); //1ページに表示するレコードの数
 
-		//一覧画面を表示
-		forward(ForwardConst.FW_EMP_INDEX);
+            //セッションにフラッシュメッセージが設定されている場合はリクエストスコープに移し替え、セッションからは削除する
+            String flush = getSessionScope(AttributeConst.FLUSH);
+            if (flush != null) {
+                putRequestScope(AttributeConst.FLUSH, flush);
+                removeSessionScope(AttributeConst.FLUSH);
+            }
 
-	}
+            //一覧画面を表示
+            forward(ForwardConst.FW_EMP_INDEX);
 
-	/**
-	 * 新規登録画面の表示
-	 * @throws ServletException
-	 * @throws IOException
-	 */
+        } //追記
 
-	public void entryNew() throws ServletException, IOException {
+    }
 
-		putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
-		putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView()); //空の従業員インスタンス
+    /**
+     * 新規登録画面を表示する
+     * @throws ServletException
+     * @throws IOException
+     */
+    public void entryNew() throws ServletException, IOException {
 
-		//新規登録画面を表示
-		forward(ForwardConst.FW_EMP_NEW);
-	}
+        //管理者かどうかのチェック //追記
+        if (checkAdmin()) { //追記
+            putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
+            putRequestScope(AttributeConst.EMPLOYEE, new EmployeeView()); //空の従業員インスタンス
 
-	/**
-	 * 新規登録を行う
-	 * @throws ServletException
-	 * @throws IOException
-	 */
-	public void create() throws ServletException, IOException {
-		//CSRF対策 token のチェック
-		if (checkToken()) {
+            //新規登録画面を表示
+            forward(ForwardConst.FW_EMP_NEW);
+        } //追記
+    }
 
-			//パラメータの値を基に従業員情報のインスタンスを作成する
-			EmployeeView ev = new EmployeeView(
-					null,
-					getRequestParam(AttributeConst.EMP_CODE),
-					getRequestParam(AttributeConst.EMP_NAME),
-					getRequestParam(AttributeConst.EMP_PASS),
-					toNumber(getRequestParam(AttributeConst.EMP_ADMIN_FLG)),
-					null,
-					null,
-					AttributeConst.DEL_FLAG_FALSE.getIntegerValue());
-			//アプリケーションスコープからpepper文字列取得
-			String pepper = getContextScope(PropertyConst.PEPPER);
+    /**
+     * 新規登録を行う
+     * @throws ServletException
+     * @throws IOException
+     */
+    public void create() throws ServletException, IOException {
 
-			//従業員情報登録
-			List<String> errors = service.create(ev, pepper);
+        //CSRF対策 tokenのチェック
+        if (checkAdmin() && checkToken()) { //追記
 
-			if (errors.size() > 0) {
-				//登録中にエラー
-				putRequestScope(AttributeConst.TOKEN, getTokenId());//CSRF対策トークン
-				putRequestScope(AttributeConst.EMPLOYEE, ev);//入力された従業員情報
-				putRequestScope(AttributeConst.ERR, errors);
+            //パラメータの値を元に従業員情報のインスタンスを作成する
+            EmployeeView ev = new EmployeeView(
+                    null,
+                    getRequestParam(AttributeConst.EMP_CODE),
+                    getRequestParam(AttributeConst.EMP_NAME),
+                    getRequestParam(AttributeConst.EMP_PASS),
+                    toNumber(getRequestParam(AttributeConst.EMP_ADMIN_FLG)),
+                    null,
+                    null,
+                    AttributeConst.DEL_FLAG_FALSE.getIntegerValue());
 
-				//新規登録画面を再表示
-				forward(ForwardConst.FW_EMP_NEW);
-			} else {
-				//登録中にエラーが無かった場合
+            //アプリケーションスコープからpepper文字列を取得
+            String pepper = getContextScope(PropertyConst.PEPPER);
 
-				//セッション中に登録完了のフラッシュメッセージを設定
-				putSessionScope(AttributeConst.FLUSH, MessageConst.I_REGISTERED.getMessage());
+            //従業員情報登録
+            List<String> errors = service.create(ev, pepper);
 
-				//一覧画面にリダイレクト
-				redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
-			}
+            if (errors.size() > 0) {
+                //登録中にエラーがあった場合
 
-		}
+                putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
+                putRequestScope(AttributeConst.EMPLOYEE, ev); //入力された従業員情報
+                putRequestScope(AttributeConst.ERR, errors); //エラーのリスト
 
-	}
+                //新規登録画面を再表示
+                forward(ForwardConst.FW_EMP_NEW);
 
-	/**
-	 * 詳細画面を表示
-	 * @throws ServletException
-	 * @throws IOException
-	 */
-	public void show() throws ServletException, IOException {
+            } else {
+                //登録中にエラーがなかった場合
 
-		//idを条件に授業員データを取得する
-		EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
-		if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
+                //セッションに登録完了のフラッシュメッセージを設定
+                putSessionScope(AttributeConst.FLUSH, MessageConst.I_REGISTERED.getMessage());
 
-			//データを削除または、論理削除されている場合のエラー画面表示
-			forward(ForwardConst.FW_ERR_UNKNOWN);
-			return;
-		}
+                //一覧画面にリダイレクト
+                redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
+            }
 
-		putRequestScope(AttributeConst.EMPLOYEE, ev);//取得した従業員情報
-		//詳細画面を表示
-		forward(ForwardConst.FW_EMP_SHOW);
-	}
+        }
+    }
 
-	/**
+    /**
+     * 詳細画面を表示する
+     * @throws ServletException
+     * @throws IOException
+     */
+    public void show() throws ServletException, IOException {
+
+        //管理者かどうかのチェック //追記
+        if (checkAdmin()) { //追記
+
+            //idを条件に従業員データを取得する
+            EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
+
+            if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
+
+                //データが取得できなかった、または論理削除されている場合はエラー画面を表示
+                forward(ForwardConst.FW_ERR_UNKNOWN);
+                return;
+            }
+
+            putRequestScope(AttributeConst.EMPLOYEE, ev); //取得した従業員情報
+
+            //詳細画面を表示
+            forward(ForwardConst.FW_EMP_SHOW);
+        } //追記
+
+    }
+
+    /**
      * 編集画面を表示する
      * @throws ServletException
      * @throws IOException
      */
-    public void edit() throws ServletException, IOException{
-    	//idを条件に授業員データを取得
-    	EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
-    	if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
-    		//データが取得できなかった、または論理削除されている場合はエラー画面を表示
-    		forward(ForwardConst.FW_ERR_UNKNOWN);
-    		return;
-    	}
+    public void edit() throws ServletException, IOException {
 
-    	putRequestScope(AttributeConst.TOKEN, getTokenId());//CSRF対策用トークン
-    	putRequestScope(AttributeConst.EMPLOYEE, ev);//取得した従業員情報
+        //管理者かどうかのチェック //追記
+        if (checkAdmin()) { //追記
 
-    	//編集画面を表示
-    	forward(ForwardConst.FW_EMP_EDIT);
+            //idを条件に従業員データを取得する
+            EmployeeView ev = service.findOne(toNumber(getRequestParam(AttributeConst.EMP_ID)));
+
+            if (ev == null || ev.getDeleteFlag() == AttributeConst.DEL_FLAG_TRUE.getIntegerValue()) {
+
+                //データが取得できなかった、または論理削除されている場合はエラー画面を表示
+                forward(ForwardConst.FW_ERR_UNKNOWN);
+                return;
+            }
+
+            putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
+            putRequestScope(AttributeConst.EMPLOYEE, ev); //取得した従業員情報
+
+            //編集画面を表示する
+            forward(ForwardConst.FW_EMP_EDIT);
+
+        } //追記
     }
 
-	/**
+    /**
      * 更新を行う
      * @throws ServletException
      * @throws IOException
@@ -178,43 +203,43 @@ public class EmployeeAction extends ActionBase {
     public void update() throws ServletException, IOException {
 
         //CSRF対策 tokenのチェック
-        if (checkToken()) {
+        if (checkAdmin() && checkToken()) { //追記
             //パラメータの値を元に従業員情報のインスタンスを作成する
-		EmployeeView ev = new EmployeeView(
-				toNumber(getRequestParam(AttributeConst.EMP_ID)),
-				getRequestParam(AttributeConst.EMP_CODE),
-				getRequestParam(AttributeConst.EMP_NAME),
-				getRequestParam(AttributeConst.EMP_PASS),
-				toNumber(getRequestParam(AttributeConst.EMP_ADMIN_FLG)),
-				null,
-				null,
-				AttributeConst.DEL_FLAG_FALSE.getIntegerValue()
-				);
-		//アプリケーションスコープからpepper文字列取得
-		String pepper = getContextScope(PropertyConst.PEPPER);
+            EmployeeView ev = new EmployeeView(
+                    toNumber(getRequestParam(AttributeConst.EMP_ID)),
+                    getRequestParam(AttributeConst.EMP_CODE),
+                    getRequestParam(AttributeConst.EMP_NAME),
+                    getRequestParam(AttributeConst.EMP_PASS),
+                    toNumber(getRequestParam(AttributeConst.EMP_ADMIN_FLG)),
+                    null,
+                    null,
+                    AttributeConst.DEL_FLAG_FALSE.getIntegerValue());
 
-		//従業員情報更新
-		List<String> errors = service.update(ev, pepper);
+            //アプリケーションスコープからpepper文字列を取得
+            String pepper = getContextScope(PropertyConst.PEPPER);
 
-		if (errors.size() > 0) {
-			//更新中にエラーが発生した場合
-			putRequestScope(AttributeConst.TOKEN, getTokenId());
-			putRequestScope(AttributeConst.EMPLOYEE, ev);
-			putRequestScope(AttributeConst.ERR, errors);
+            //従業員情報更新
+            List<String> errors = service.update(ev, pepper);
 
-			//編集画面を再表示
-			forward(ForwardConst.FW_EMP_EDIT);
+            if (errors.size() > 0) {
+                //更新中にエラーが発生した場合
 
-		} else {
-			//更新中にエラーが無かった場合
-			//セッションに更新完了のフラッシュメッセージを設定
-			putSessionScope(AttributeConst.FLUSH, MessageConst.I_UPDATED.getMessage());
-			//一覧画面にリダイレクト
-			redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
-		}
+                putRequestScope(AttributeConst.TOKEN, getTokenId()); //CSRF対策用トークン
+                putRequestScope(AttributeConst.EMPLOYEE, ev); //入力された従業員情報
+                putRequestScope(AttributeConst.ERR, errors); //エラーのリスト
 
-       }
+                //編集画面を再表示
+                forward(ForwardConst.FW_EMP_EDIT);
+            } else {
+                //更新中にエラーがなかった場合
 
+                //セッションに更新完了のフラッシュメッセージを設定
+                putSessionScope(AttributeConst.FLUSH, MessageConst.I_UPDATED.getMessage());
+
+                //一覧画面にリダイレクト
+                redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
+            }
+        }
     }
 
     /**
@@ -223,34 +248,43 @@ public class EmployeeAction extends ActionBase {
      * @throws IOException
      */
     public void destroy() throws ServletException, IOException {
-    	//CSRF対策 token のチェック
-    	if (checkToken()) {
-    		//idを条件に従業員データを論理削除
-    		service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
 
-    		//セッションに削除官僚のフラッシュメッセージを設定
-    		putSessionScope(AttributeConst.FLUSH, MessageConst.I_DELETED.getMessage());
+        //CSRF対策 tokenのチェック
+        if (checkAdmin() && checkToken()) { //追記
 
-    		//一覧画面にリダイレクト
-    		redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
-    	}
+            //idを条件に従業員データを論理削除する
+            service.destroy(toNumber(getRequestParam(AttributeConst.EMP_ID)));
+
+            //セッションに削除完了のフラッシュメッセージを設定
+            putSessionScope(AttributeConst.FLUSH, MessageConst.I_DELETED.getMessage());
+
+            //一覧画面にリダイレクト
+            redirect(ForwardConst.ACT_EMP, ForwardConst.CMD_INDEX);
+        }
     }
 
+    /**
+     * ログイン中の従業員が管理者かどうかチェックし、管理者でなければエラー画面を表示
+     * true: 管理者 false: 管理者ではない
+     * @throws ServletException
+     * @throws IOException
+     */
+    private boolean checkAdmin() throws ServletException, IOException {
 
+        //セッションからログイン中の従業員情報を取得
+        EmployeeView ev = (EmployeeView) getSessionScope(AttributeConst.LOGIN_EMP);
 
+        //管理者でなければエラー画面を表示
+        if (ev.getAdminFlag() != AttributeConst.ROLE_ADMIN.getIntegerValue()) {
 
+            forward(ForwardConst.FW_ERR_UNKNOWN);
+            return false;
 
+        } else {
 
+            return true;
+        }
 
-
-
-
-
-
-
-
-
-
-
+    }
 
 }
